misc/numfun.h: fix warning

+++ misc/numfun.h
@@ -29,7 +29,7 @@
 		if (v == 0) return 0; \
 		t a = v | (v - 1) | (v >> 1); \
 		t b = v << 1; \
-		for (int i = 0; i < ((sizeof(t) * CHAR_BIT) - 3); i++) { \
+		for (unsigned long i = 0; i < ((sizeof(t) * CHAR_BIT) - 3); i++) { \
 			a = (a | b) >> 1; \
 		} \
 		return a; \
